refactor: Updated message-shim to remove cognitive complexity violations.
 Fixed jsdoc warnings as well.
