+++ lib/shim/specs/index.js
@@ -37,6 +37,8 @@ exports.MiddlewareSpec = MiddlewareSpec
 exports.RecorderSpec = RecorderSpec
 exports.SegmentSpec = SegmentSpec
 exports.WrapSpec = WrapSpec
+exports.MessageSpec = MessageSpec
+exports.MessageSubscribeSpec = MessageSubscribeSpec
 
 function cast(Class, spec) {
   return spec instanceof Class ? spec : new Class(spec)
@@ -79,6 +81,24 @@ function MiddlewareSpec(spec) {
   this.appendPath = hasOwnProperty(spec, 'appendPath') ? spec.appendPath : true
 }
 util.inherits(MiddlewareSpec, RecorderSpec)
+
+function MessageSpec(spec) {
+  RecorderSpec.call(this, spec)
+  this.destinationName = hasOwnProperty(spec, 'destinationName') ? spec.destinationName : null
+  this.destinationType = hasOwnProperty(spec, 'destinationType') ? spec.destinationType : null
+  this.headers = hasOwnProperty(spec, 'headers') ? spec.headers : null
+  this.routingKey = hasOwnProperty(spec, 'routingKey') ? spec.routingKey : null
+  this.queue = hasOwnProperty(spec, 'queue') ? spec.queue : null
+  this.messageHandler = hasOwnProperty(spec, 'messageHandler') ? spec.messageHandler : null
+}
+
+util.inherits(MessageSpec, RecorderSpec)
+
+function MessageSubscribeSpec(spec) {
+  MessageSpec.call(this, spec)
+  this.consumer = hasOwnProperty(spec, 'consumer') ? spec.consumer : null
+}
+util.inherits(MessageSubscribeSpec, MessageSpec)
 
 function _defaultGetParams(shim, fn, name, args, req) {
   return req && req.params
